Document cloud-agent tracing authentication
