+++ app/src/tracing.rs
@@ -4,7 +4,7 @@ use std::time::Duration;
 use tracing::subscriber;
 
 #[cfg(not(target_family = "wasm"))]
-mod auth;
+mod cloud_agent_auth;
 #[cfg(not(target_family = "wasm"))]
 mod native;
 
@@ -23,6 +23,11 @@ pub fn init() -> anyhow::Result<Initialization> {
 }
 
 #[cfg(not(target_family = "wasm"))]
+/// Starts cloud-agent trace credential refresh after authenticated application services exist.
+///
+/// The exporter and dispatch credential are initialized earlier by [`init`]. This later lifecycle
+/// hook supplies the authenticated managed-secrets client needed to mint replacements without
+/// broadening tracing initialization to ordinary application processes.
 pub fn start_auth_refresh(
     client: std::sync::Arc<dyn warp_managed_secrets::client::ManagedSecretsClient>,
     ctx: &mut warpui::AppContext,
